warehouse_api/requirements.txt
Flask==2.0.1
requests
mysql-connector